feat: handle like and trash buttons

+++ src/utils/Api.js
@@ -52,20 +52,20 @@ class Api{
     .then(res => this._checkApiResponse(res))
   }
 
-  putCardLike(cardId) {
-    return fetch(`${this._host}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-  })
-    .then(res => this._checkApiResponse(res))
-  }
-
-  removeCardLike(cardId) {
+  changeLikeCardStatus(cardId, IsLiked) {
+    if (IsLiked) {
+      return fetch(`${this._host}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers
+    })
+      .then(res => this._checkApiResponse(res))
+    } 
     return fetch(`${this._host}/cards/${cardId}/likes`, {
       method: 'DELETE',
       headers: this._headers
-  })
-    .then(res => this._checkApiResponse(res))
+    })
+      .then(res => this._checkApiResponse(res))
+    
   }
 
   deleteCard(cardId) {
